add guice, and refactor RootProcess
