+++ app/build.gradle
@@ -24,4 +24,5 @@ dependencies {
   testCompile 'junit:junit:4.12'
   compile 'com.android.support:appcompat-v7:23.1.1'
   compile 'com.google.android.gms:play-services-analytics:8.4.0'
+  compile 'com.google.inject:guice:2.0-no_aop'
 }
