+++ app/src/main/java/org/shen/xi/resetwifi/MainApplication.java
@@ -8,7 +8,6 @@ import com.google.android.gms.analytics.Tracker;
 
 public class MainApplication extends Application {
 
-
   private GoogleAnalytics googleAnalytics;
 
   public void initializeContext(Context context) {
