+++ app/src/main/java/org/shen/xi/resetwifi/MainModule.java
@@ -1,0 +1,14 @@
+package org.shen.xi.resetwifi;
+
+import com.google.inject.AbstractModule;
+
+public class MainModule extends AbstractModule {
+  @Override
+  protected void configure() {
+    if (BuildConfig.DEBUG) {
+      bind(RootProcess.class).to(FakeRootProcess.class);
+    } else {
+      bind(RootProcess.class).to(SuRootProcess.class);
+    }
+  }
+}
